+++ RELEASES.md
@@ -4,6 +4,7 @@
 
 Notable changes:
  - Use proto3 JSON mapping for web requests: https://developers.google.com/protocol-buffers/docs/proto3#json
+ - Java indexer: report error when indexing from compilation's source root
 
 ## v0.0.15